functions/requirements.txt
firebase_functions~=0.1.0
fastapi~=0.109.0
asgiref~=3.7.2
requests~=2.32.0
google-cloud-pubsub~=2.25.0
./frcm-0.1.0-py3-none-any.whl
